+++ src/solver/conv_asm_implicit_gemm_gtc_fwd.cpp
@@ -1569,7 +1569,7 @@ ConvAsmImplicitGemmGTCDynamicFwdXdlops::GetSolution(const ExecutionContext& ctx,
 
     kernel.kernel_name = kernel_name;
     kernel.g_wk.clear();
-    kernel.g_wk.push_back(grid_size * block_size);
+    kernel.g_wk.push_back(static_cast<std::size_t>(grid_size) * block_size);
     kernel.g_wk.push_back(1);
     kernel.g_wk.push_back(1);
     kernel.l_wk.clear();
